Move tests to __tests__ folder and revert peopleapp changes
- Move AI SDK adapter tests into src/ai-sdk/__tests__/
- Revert peopleapp to base branch state (remove ai/ai-sdk deps and
  the AI SDK provider usage demo)
